requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
astroid==3.2.4
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
distlib==0.3.8
Django==5.0.2
django-filter==24.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
passlib==1.7.4
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zope.interface==7.0.3
